requirements.txt
Pillow==9.1.0
PyMuPDF==1.19.6

